implement passport jwt

+++ controllers/api/v1/post_api.js
@@ -1,24 +1,29 @@
 const Post = require('../../../models/post');
 const Comment = require('../../../models/comment');
 
-module.exports.destroy = async(request,response)=>{
-    try{
+module.exports.destroy = async (request, response) => {
+    try {
         let post = await Post.findById(request.params.id);
+        if (post.user == request.user.id) {
+            Post.deleteOne(post).then(function () {
+                console.log('DELETE THE POST');
+            }).catch(function () {
+                console.log('Life is hard');
+            })
 
-        Post.deleteOne(post).then(function(){
-            console.log('DELETE THE POST');
-        }).catch(function(){
-            console.log('Life is hard');
-        })
-
-        await Comment.deleteMany({post:request.params.id});
-        return response.status(200).json({
-            message:'Post Deleted Success'
-        })
-    }catch(error){
+            await Comment.deleteMany({ post: request.params.id });
+            return response.status(200).json({
+                message: 'Post Deleted Success'
+            })
+        } else {
+            return response.status(401).json({
+                message:'You cannot delete the post'
+            })
+        }
+    } catch (error) {
         console.log(error);
         return response.status(500).json({
-            message:error
+            message: error
         });
     }
 }
@@ -26,14 +31,14 @@ module.exports.destroy = async(request,response)=>{
 module.exports.index = async (request, response) => {
 
     let posts = await Post.find({})
-    .sort('-createdAt')
-    .populate('user')
-    .populate({
-        path: 'comments',
-        populate: {
-            path: 'user'
-        }
-    });
+        .sort('-createdAt')
+        .populate('user')
+        .populate({
+            path: 'comments',
+            populate: {
+                path: 'user'
+            }
+        });
